requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-money==2.1.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
httpie==3.0.2
idna==3.3
ipykernel==6.9.1
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
py-moneyed==1.2
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
